requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
appdirs==1.4.4
    # via requests-cache
attrs==21.4.0
    # via
    #   cattrs
    #   requests-cache
cattrs==22.1.0
    # via requests-cache
certifi==2022.6.15.1
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.1
    # via requests
cryptography==37.0.4
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
exceptiongroup==1.0.0rc8
    # via
    #   cattrs
    #   requests-cache
idna==3.3
    # via requests
jsonref==0.2
    # via -r requirements.in
mergedeep==1.3.4
    # via -r requirements.in
openpyxl==3.0.10
    # via -r requirements.in
polling==0.3.2
    # via -r requirements.in
pycparser==2.21
    # via cffi
pyjwt==2.4.0
    # via -r requirements.in
pyyaml==6.0
    # via -r requirements.in
requests[security]==2.28.1
    # via
    #   -r requirements.in
    #   requests-cache
requests-cache==0.9.5
    # via -r requirements.in
six==1.16.0
    # via url-normalize
url-normalize==1.4.3
    # via requests-cache
urllib3==1.26.11
    # via
    #   requests
    #   requests-cache
xlsxwriter==3.0.3
    # via -r requirements.in
